+++ policyengine_us/variables/gov/irs/credits/recovery_rebate_credit/recovery_rebate_credit.py
@@ -9,10 +9,4 @@ class recovery_rebate_credit(Variable):
     definition_period = YEAR
     reference = "https://www.law.cornell.edu/uscode/text/26/6428"
 
-    formula = sum_of_variables(
-        [
-            "rrc_cares",
-            "rrc_caa",
-            "rrc_arpa",
-        ]
-    )
+    adds = ["rrc_cares", "rrc_caa", "rrc_arpa"]