+++ policyengine_us/variables/gov/irs/income/taxable_income/deductions/standard_deduction/standard_deduction.py
@@ -9,6 +9,4 @@ class standard_deduction(Variable):
     definition_period = YEAR
     reference = "https://www.law.cornell.edu/uscode/text/26/63#c"
 
-    formula = sum_of_variables(
-        ["basic_standard_deduction", "additional_standard_deduction"]
-    )
+    adds = ["basic_standard_deduction", "additional_standard_deduction"]