+++ policyengine_us/variables/gov/states/in/tax/income/add_backs/in_add_backs.py
@@ -11,13 +11,11 @@ class in_add_backs(Variable):
         "http://iga.in.gov/legislative/laws/2021/ic/titles/006#6-3-1-3.5"
     )
 
-    formula = sum_of_variables(
-        [
-            "in_bonus_depreciation_add_back",
-            "in_nol_add_back",
-            "in_oos_municipal_obligation_interest_add_back",
-            "in_other_add_backs",
-            "in_section_179_expense_add_back",
-            "in_tax_add_back",
-        ]
-    )
+    adds = [
+        "in_bonus_depreciation_add_back",
+        "in_nol_add_back",
+        "in_oos_municipal_obligation_interest_add_back",
+        "in_other_add_backs",
+        "in_section_179_expense_add_back",
+        "in_tax_add_back",
+    ]