+++ policyengine_us/variables/gov/states/in/tax/income/exemptions/in_exemptions.py
@@ -10,11 +10,9 @@ class in_exemptions(Variable):
     reference = "http://iga.in.gov/legislative/laws/2021/ic/titles/006#6-3-1-3.5"  # (a)(1)-(5)
     defined_for = StateCode.IN
 
-    formula = sum_of_variables(
-        [
-            "in_base_exemptions",
-            "in_additional_exemptions",
-            "in_aged_blind_exemptions",
-            "in_aged_low_agi_exemptions",
-        ]
-    )
+    adds = [
+        "in_base_exemptions",
+        "in_additional_exemptions",
+        "in_aged_blind_exemptions",
+        "in_aged_low_agi_exemptions",
+    ]