+++ policyengine_us/variables/gov/states/ut/tax/income/credits/taxpayer_credit/ut_taxpayer_credit.py
@@ -9,9 +9,5 @@ class ut_taxpayer_credit(Variable):
     definition_period = YEAR
     defined_for = StateCode.UT
 
-    def formula(tax_unit, period, parameters):
-        ut_taxpayer_credit_max = tax_unit("ut_taxpayer_credit_max", period)
-        ut_taxpayer_credit_reduction = tax_unit(
-            "ut_taxpayer_credit_reduction", period
-        )
-        return ut_taxpayer_credit_max - ut_taxpayer_credit_reduction
+    adds = ["ut_taxpayer_credit_max"]
+    subtracts = ["ut_taxpayer_credit_reduction"]