+++ policyengine_us/variables/household/expense/housing/housing_cost.py
@@ -8,4 +8,4 @@ class housing_cost(Variable):
     unit = USD
     definition_period = YEAR
 
-    formula = sum_of_variables(["rent"])
+    adds = ["rent"]