+++ policyengine_us/variables/household/income/spm_unit/spm_unit_taxes.py
@@ -8,12 +8,10 @@ class spm_unit_taxes(Variable):
     definition_period = YEAR
     unit = USD
 
-    formula = sum_of_variables(
-        [
-            "spm_unit_payroll_tax",
-            "spm_unit_self_employment_tax",
-            "spm_unit_federal_tax",
-            "spm_unit_state_tax",
-            "flat_tax",
-        ]
-    )
+    adds = [
+        "spm_unit_payroll_tax",
+        "spm_unit_self_employment_tax",
+        "spm_unit_federal_tax",
+        "spm_unit_state_tax",
+        "flat_tax",
+    ]