+++ CHANGELOG.md
@@ -5,6 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.199.1] - 2023-01-10 12:30:48
+
+### Changed
+
+- Used adds and subtracts instead of sum_of_variables and Python code.
+
+## [0.199.0] - 2023-01-10 06:45:38
+
+### Added
+
+- OH Income Tax Rates
+
+## [0.198.0] - 2023-01-09 16:06:12
+
+### Added
+
+- MD AGI subtraction for pension income.
+
+## [0.197.7] - 2023-01-06 11:26:07
+
+### Changed
+
+- Core bumped to 1.11.1.
+
 ## [0.197.6] - 2023-01-06 01:56:20
 
 ### Fixed
@@ -2529,6 +2553,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+[0.199.1]: https://github.com/PolicyEngine/policyengine-us/compare/0.199.0...0.199.1
+[0.199.0]: https://github.com/PolicyEngine/policyengine-us/compare/0.198.0...0.199.0
+[0.198.0]: https://github.com/PolicyEngine/policyengine-us/compare/0.197.7...0.198.0
+[0.197.7]: https://github.com/PolicyEngine/policyengine-us/compare/0.197.6...0.197.7
 [0.197.6]: https://github.com/PolicyEngine/policyengine-us/compare/0.197.5...0.197.6
 [0.197.5]: https://github.com/PolicyEngine/policyengine-us/compare/0.197.4...0.197.5
 [0.197.4]: https://github.com/PolicyEngine/policyengine-us/compare/0.197.3...0.197.4